Moving to Dynamic Volume Clouds
Still WIP. The idea is that the clouds will be:
1) More procedural
2) More varied
3) MUCH more versatile
4) Better performance

Step 1 is complete. The particles are re-aligned to the camera and move
in 1-unit increments as the camera moves. This is a single
check/movement, so it results in VERY good performance.

Next up is modifying the shader to vary size/rotation based off of
position.