Adds UVNoise for clouds.
The new properties of the EVE_CLOUDS object items are:

_UVNoiseTex - a UV displacement texture using RG channels.
_UVNoiseScale - the size of the fractional part of the UV coord used for indexing the noise texture. Should be 1/W where W is a whole number to avoid edge stitching problems.
_UVNoiseStrength - the amount of displacement. Should generally be less than _UVNoiseScale.
_UVNoiseAnimation - offset multiplier for _Time allowing subtle animation of the noise by traversing the noise texture.

clouds.cfg is tweaked to make the most of the new options.

City Lights do not use the feature, but they could be changed to (without animation of course, and probably with a different noise texture).

The _DetailTex can generally be scaled to appear larger (smaller _DetailScale config value) as it is no longer needed for obscuring pixelization.