Fix Issue #4 - Artifacts around landing gear - Shadows.
Should also fix other transparent objects.
Third-party mods with similar problems with semi-transparent materials should ensure such objects are not on one of the shadowed layers, which are:
            Default = 0
            Water = 4
            Kerbals = 16
            Parts = 19
In particular, transparent lighting effects should be on layer:
            TransparentFX = 1
It is however perfectly fine for "solid" transparent objects such as glass to be shadowed; just be aware that the entire geometry gets shadow applied, not just the opaque areas.